// src/events/dto/enum.ts
export enum EventType {
  GENDER_SPECIFIC = 'gender_specific', // Evento com vagas divididas por gênero
  GENERAL = 'general', // Evento sem divisão por gênero
}

export enum UserType {
  CLIENT = 'client', // Cliente
  STAFF = 'staff', // Membro da equipe
}

export enum Gender {
  MALE = 'male', // Masculino
  FEMALE = 'female',
}